refactor(web): extract task-create-dialog setup hooks to keep file under 600 lines
Pull TaskCreateDialogProps interface and the dialog setup hooks
(useEnhanceForDialog, useJiraImportHandler, useLinearImportHandler,
useSubmitHandlersWiring, useTaskCreateDialogSetup) out of
task-create-dialog.tsx into a dedicated task-create-dialog-setup.tsx.
The Improve Kandev slot props pushed the main file over the 600-line
ESLint max-lines limit; the extraction restores compliance without
behavior changes.
